refactor: modified engine's SetPlayer related stuff

+++ Assets/Scripts/Interfaces/IEngine.cs
@@ -3,7 +3,7 @@ interface IEngine
     public void PlayManual(int[] move);
     public int[] PlayAuto(int recursionDepth);
     public void SetState(byte[] newState);
-    public void SetPlayer(int colour);
+    public void SetPlayer(byte colour);
     public float Evaluate();
     public bool CheckWin();
 }
